+++ src/rest-api/errors.ts
@@ -1,5 +1,5 @@
-/** Pic wasn't a PNG or JPEG not exceeding 3 MiB. */
-export class InvalidPicError extends Error {}
+/** The image wasn't a PNG or JPEG not exceeding 3 MiB. */
+export class InvalidImageError extends Error {}
 
 /** Audio wasn't an MP3/MP4 not exceeding 3 MiB. */
 export class InvalidAudioError extends Error {}