+++ CHANGELOG.md
@@ -5,6 +5,39 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.0](https://github.com/neelkamath/omni-chat-js/releases/tag/v3.0.0) - 2021-07-15
+
+### Added
+
+- `interface ImageFile`
+- `interface AudioFile`
+- `interface VideoFile`
+- `interface DocFile`
+
+### Changed
+
+- Rename `class InvalidPicError` to `class InvalidImageError`.
+- Rename `type PicType` to `type ImageType`.
+- Rename `type Pic` to `type Image`.
+- Rename `getProfilePic()` to `getProfileImage()`.
+- Rename `patchProfilePic()` to `patchProfileImage()`.
+- Rename `getGroupChatPic()` to `getGroupChatImage()`.
+- Remove the `accessToken` parameter from `getGroupChatImage()`.
+- Rename `patchGroupChatPic()` to `patchGroupChatImage()`.
+- Rename `getPicMessage()` to `getImageMessage()`.
+- Rename `postPicMessage()` to `postImageMessage()`.
+- Return an `interface ImageFile` instead of a `type Image` from `getProfileImage()`, `getGroupChatImage()`, and `getImageMessage()`.
+- Return an `interface AudioFile` instead of a `type Audio` from `getAudioMessage()`.
+- Return an `interface VideoFile` instead of a `type Video` from `getVideoMessage()`.
+- Return an `interface DocFile` instead of a `type Doc` from `getDocMessage()`.
+
+### Removed
+
+- Remove `type Image` in favor of `interface ImageFile`.
+- Remove `type Audio` in favor of `interface AudioFile`.
+- Remove `type Video` in favor of `interface VideoFile`.
+- Remove `type Doc` in favor of `interface DocFile`.
+
 ## [2.0.1](https://github.com/neelkamath/omni-chat-js/releases/tag/v2.0.1) - 2021-07-07
 
 ### Fixed